requirements.txt
absl-py==0.10.0
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.21
detectron2==0.2.1+cu101
future==0.18.2
fvcore==0.1.2.post20201020
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.33.1
idna==2.10
kiwisolver==1.2.0
Markdown==3.3.2
matplotlib==3.3.2
mock==4.0.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
Pillow==8.0.0
portalocker==2.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pydot==1.4.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
urllib3==1.25.11
Werkzeug==1.0.1
yacs==0.1.8
